+++ src/components/MapRenderer/CategoricalPictogramMapRenderer.tsx
@@ -20,21 +20,30 @@ const directives = [ // eslint-disable-line @typescript-eslint/no-unused-vars
   bindData,
 ];
 
+type CategoricalPictoMapK = string | number | null | undefined;
+type CategoricalPictoMapV = [string, string, [number, number]] | null;
+
 export default function categoricalPictogramRenderer(
   layerDescription: LayerDescriptionCategoricalPictogram,
 ): JSX.Element {
   let refElement: SVGGElement;
 
   const symbolMap = createMemo(
-    () => new Map<string | number | null | undefined, [string, string, [number, number]]>(
-      layerDescription.rendererParameters.mapping
-        .map(({
-          value,
-          iconContent,
-          iconType,
-          iconDimension,
-        }) => [value, [iconType, iconContent, iconDimension]]),
-    ),
+    () => {
+      const map = new Map<CategoricalPictoMapK, CategoricalPictoMapV>(
+        layerDescription.rendererParameters.mapping
+          .map(({
+            value,
+            iconContent,
+            iconType,
+            iconDimension,
+          }) => [value, [iconType, iconContent, iconDimension]]),
+      );
+      map.set('', null);
+      map.set(null, null);
+      map.set(undefined, null);
+      return map;
+    },
   );
 
   return <g
@@ -61,28 +70,28 @@ export default function categoricalPictogramRenderer(
         (feature) => {
           const icon = createMemo(() => symbolMap()
             .get(feature.properties[layerDescription.rendererParameters.variable]));
+          if (!icon()) return <></>;
           const projectedCoords = createMemo(
             () => globalStore.projection(feature.geometry.coordinates)
-              .map((d, i) => d - icon()[2][i] / 2),
+              .map((d, i) => d - icon()![2][i] / 2),
           );
-          if (!icon()) return <></>;
-          if (icon()[0] === 'SVG') {
+          if (icon()![0] === 'SVG') {
             return <g
               transform={`translate(${projectedCoords()[0]}, ${projectedCoords()[1]})`}
-              mgt:icon-dimension={JSON.stringify(icon()[2])}
+              mgt:icon-dimension={JSON.stringify(icon()![2])}
               // @ts-expect-error because use:bind-data isn't a property of this element
               use:bindData={feature}
               // eslint-disable-next-line solid/no-innerhtml
-              innerHTML={setWidthHeight(icon()[1], icon()[2][0], icon()[2][1])}
+              innerHTML={setWidthHeight(icon()![1], icon()![2][0], icon()![2][1])}
             />;
           }
           return <g
             transform={`translate(${projectedCoords()[0]}, ${projectedCoords()[1]})`}
-            mgt:icon-dimension={JSON.stringify(icon()[2])}
+            mgt:icon-dimension={JSON.stringify(icon()![2])}
             // @ts-expect-error because use:bind-data isn't a property of this element
             use:bindData={feature}
           >
-            <image width={icon()[2][0]} height={icon()[2][1]} href={icon()[1]}/>
+            <image width={icon()![2][0]} height={icon()![2][1]} href={icon()![1]}/>
           </g>;
         }
       }